# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.0",
    "title": "오픈마켓 API",
    "description": "오픈마켓 API Server입니다.<br>\n      <a href=\"/\">버전별 변경사항 확인</a>"
  },
  "servers": [
    {
      "url": "https://api.fesp.shop",
      "description": ""
    },
    {
      "url": "http://localhost",
      "description": "로컬 테스트"
    }
  ],
...
  "paths": {
    "/users/": {
      "post": {
        "tags": [
          "회원"
        ],
        "summary": "회원 가입",
        "description": "회원 가입을 합니다.<br>회원 가입을 완료한 후 회원 정보를 반환합니다.",
        "responses": {
          "201": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "기본 속성": {
                    "$ref": "#/components/examples/createUserRes"
                  },
                  "extra 속성": {
                    "$ref": "#/components/examples/createUserResWithExtra"
                  }
                }
              }
            }
          },
          "409": {
            "description": "이메일 중복됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error409"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "requestBody": {
          "description": "회원 정보가 저장된 객체입니다.<br>email: 이메일(필수)<br>password: 비밀번호(필수)<br>name: 이름(필수)<br>phone: 전화번호(선택)<br>address: 주소(선택)<br>type: 회원 구분(필수, 구매회원: user, 판매회원: seller)<br>extra: 추가 데이터(선택). 추가하고 싶은 아무 속성이나 지정",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/createUser"
              },
              "examples": {
                "기본 속성": {
                  "$ref": "#/components/examples/createUser"
                },
                "extra 속성": {
                  "$ref": "#/components/examples/createUserWithExtra"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{_id}/{*}": {
      "get": {
        "tags": [
          "회원"
        ],
        "summary": "회원 정보 조회(단일 속성)",
        "description": "회원 정보 중 한가지 속성을 조회한다.",
        "parameters": [
          {
            "name": "_id",
            "description": "조회할 회원 id",
            "in": "path",
            "example": "5",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "*",
            "description": "조회할 속성<br>단일 속성 조회 예시: name<br>중첩 속성 조회 예시: extra/addressBook",
            "in": "path",
            "example": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "users/5/name 조회": {
                    "$ref": "#/components/examples/userInfoResOneAttr"
                  },
                  "users/5/extra/addressBook 조회": {
                    "$ref": "#/components/examples/userInfoResWithExtra"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
    "/users/{_id}": {
      "get": {
        "tags": [
          "회원"
        ],
        "summary": "회원 정보 조회(모든 속성)",
        "description": "회원 정보의 모든 속성을 조회합니다.<br>\r  등록한 속성 이외에 다음의 정보가 추가됩니다.<br>\r  posts: 지정한 사용자가 작성한 게시글 수\r  postViews: 지정한 사용자가 작성한 모든 게시글 조회수\r  bookmark.products: 지정한 사용자가 북마크 한 상품 수\r  bookmark.users: 지정한 사용자가 북마크 한 사용자 수\r  bookmark.posts: 지정한 사용자가 북마크 한 게시글 수\r  ",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "조회할 회원 id",
            "example": "5"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "기본 속성": {
                    "$ref": "#/components/examples/userInfoRes"
                  },
                  "extra 속성": {
                    "$ref": "#/components/examples/userInfoResWithExtra"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "patch": {
        "tags": [
          "회원"
        ],
        "summary": "회원 정보 수정",
        "description": "회원 정보를 수정한다.",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "회원 id",
            "example": "5"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "examples": {
                  "기본 속성": {
                    "$ref": "#/components/examples/updateUserResOneAttr"
                  },
                  "extra 속성": {
                    "$ref": "#/components/examples/updateUserResWithExtra"
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "404": {
            "description": "회원이 존재하지 않거나 접근 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error404"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "수정할 회원 정보",
          "required": true,
          "content": {
            "application/json": {
              "examples": {
                "기본 속성": {
                  "$ref": "#/components/examples/updateUserOneAttr"
                },
                "extra 속성": {
                  "$ref": "#/components/examples/updateUserWithExtra"
                }
              }
            }
          }
        }
      }
    },
...
    "/carts/local": {
      "post": {
        "tags": [
          "장바구니"
        ],
        "summary": "장바구니 목록 조회(비로그인)",
        "description": "로그인 되지 않은 상태에서 장바구니 목록을 조회합니다.<br>요청 바디에 상품 정보와 수량을 전달하면 장바구니 화면 구성에 필요한 상품 정보와 가격 정보를 응답합니다.",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cartListRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "requestBody": {
          "description": "장바구니 목록 조회를 위한 상품 정보가 저장된 객체입니다.<br>products 속성은 필수이며 객체 배열입니다.<br>배열의 요소인 객체는 다음과 같은 필수 정보를 포함해야 하고 추가 속성은 자유롭게 지정하면 됩니다.<br>_id: 상품 id<br>quantity: 구매 수량",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/cartList"
              }
            }
          }
        }
      }
    },
    "/carts/": {
      "post": {
        "tags": [
          "장바구니"
        ],
        "summary": "장바구니에 상품 추가",
        "description": "장바구니에 상품을 추가합니다.<br>이미 장바구니에 추가된 상품일 경우 수량이 증가하고 새로운 상품이 추가될 경우 목록에 추가됩니다.",
        "responses": {
          "201": {
            "description": "성공<br>추가된 상품을 포함한 장바구니 목록이 반환됩니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cartCreateRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "장바구니에 추가할 상품 정보가 저장된 객체입니다.<br>다음과 같은 필수 정보를 포함해야 합니다.<br>product_id: 상품 id<br>quantity: 구매 수량",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/cartCreate"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "장바구니"
        ],
        "summary": "장바구니 목록 조회(로그인)",
        "description": "로그인 한 사용자의 장바구니 목록을 조회합니다.",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cartListLoginRes"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      },
      "delete": {
        "tags": [
          "장바구니"
        ],
        "summary": "장바구니 상품 여러건 삭제",
        "description": "장바구니 상품을 여러건 삭제합니다.",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/simpleOK"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "삭제할 장바구니 정보가 저장된 객체입니다.<br>carts: 삭제할 장바구니 id 목록(필수, 배열)",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/cartDeleteBody"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "장바구니"
        ],
        "summary": "장바구니 합치기",
        "description": "지정한 상품 목록을 장바구니에 합칩니다.<br>이미 장바구니에 추가된 상품일 경우 수량이 증가하고 새로운 상품이 추가될 경우 목록에 추가됩니다.<br>로그인 하지 않은 상태에서 장바구니에 상품을 담으면 localStorage에 저장해 두고 결제시 로그인을 하게되면 이 API를 호출해서 서버에 저장된 장바구니 상품과 localStorage에 저장된 장바구니 상품을 합칠때 사용합니다.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "성공<br>합친 상품을 포함한 장바구니 목록이 반환됩니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/cartCreateRes"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "422": {
            "description": "입력값 검증 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error422"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ],
        "requestBody": {
          "description": "장바구니 정보가 저장된 객체입니다.<br>products 속성은 필수이며 객체 배열입니다.<br>배열의 요소인 객체는 다음과 같은 필수 정보를 포함해야 합니다.<br>_id: 상품 id<br>quantity: 구매 수량",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/cartMergeBody"
              }
            }
          }
        }
      }
    },
...
    "/carts/cleanup": {
      "delete": {
        "tags": [
          "장바구니"
        ],
        "summary": "장바구니 비우기",
        "description": "장바구니를 비웁니다.",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/simpleOK"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error401"
                }
              }
            }
          },
          "500": {
            "description": "서버 에러",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error500"
                }
              }
            }
          }
        },
        "security": [
          {
            "Access Token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Access Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      },
      "Refresh Token": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "error409": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "이미 등록된 리소스입니다."
          }
        }
      },
...
      "error422": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "잘못된 입력값이 있습니다."
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "field"
                },
                "value": {
                  "type": "string",
                  "example": "swaggermarket.com"
                },
                "msg": {
                  "type": "string",
                  "example": "이메일 형식에 맞지 않습니다."
                },
                "path": {
                  "type": "string",
                  "example": "email"
                },
                "location": {
                  "type": "string",
                  "example": "body"
                }
              }
            }
          }
        }
      },
...
    },
    "examples": {
...
      "createUserRes": {
        "value": {
          "ok": 1,
          "item": {
            "email": "gd@market.com",
            "name": "GD",
            "type": "user",
            "phone": "01011112222",
            "address": "서울시 강남구 역삼동 123",
            "_id": 5,
            "createdAt": "2023.11.21 14:31:15",
            "updatedAt": "2023.11.21 14:31:15"
          }
        }
      },
      "createUserResWithExtra": {
        "value": {
          "ok": 1,
          "item": {
            "email": "gdragon@market.com",
            "name": "G드래곤",
            "type": "user",
            "phone": "01011112222",
            "address": "서울시 강남구 역삼동 123",
            "extra": {
              "gender": "extra에는 프로젝트에서 필요한 아무 속성이나",
              "age": "넣으면 됩니다.",
              "address": [
                "배열도",
                "가능하고"
              ],
              "profileImage": {
                "path ": "/files/00-sample/profile.jpg",
                "name ": "profile.jpg",
                "originalname": "사커게임.png"
              },
              "obj": {
                "hello": "객체로",
                "hi": "넣어도 됩니다."
              },
              "addressBook": [
                {
                  "name": "집",
                  "address": "서울시"
                },
                {
                  "name": "회사",
                  "address": "인천시"
                }
              ]
            },
            "_id": 5,
            "createdAt": "2023.11.21 14:33:41",
            "updatedAt": "2023.11.21 14:33:41"
          }
        }
      },
...
      "userInfoResOneAttr": {
        "value": {
          "ok": 1,
          "item": {
            "name": "GD"
          }
        }
      },
...
    }
  }
}